add DefaultSplitter
